GlobalMusicTrendPipeline_clean/backend/requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0

# Flask API Dependencies (for JioSaavn integration)
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Data Processing & Analytics
pandas==2.1.3
numpy==1.25.2
requests==2.31.0

# Streaming & Real-time
kafka-python==2.0.2
python-socketio==5.10.0

# JioSaavn API Dependencies
pyDes==2.0.1

# Development & Logging
python-dotenv==1.0.0
python-multipart==0.0.6

# Additional utilities
asyncio-mqtt==0.16.1
aiofiles==23.2.1
